Missed a cleanup item
