+++ stellar-dotnet-sdk/responses/AssetResponse.cs
@@ -76,22 +76,4 @@ namespace stellar_dotnet_sdk.responses
             Links = links;
         }
     }
-
-
-    /// <summary>
-    /// 
-    /// </summary>
-    public class AssetResponseLinks
-    {
-        /// <summary>
-        /// 
-        /// </summary>
-        [JsonProperty(PropertyName = "toml")]
-        public Link Toml { get; set; }
-
-        public AssetResponseLinks(Link toml)
-        {
-            Toml = toml;
-        }
-    }
 }
