+++ stellar-dotnet-sdk/responses/AssetResponseLinks.cs
@@ -1,0 +1,21 @@
+﻿using Newtonsoft.Json;
+
+namespace stellar_dotnet_sdk.responses
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class AssetResponseLinks
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        [JsonProperty(PropertyName = "toml")]
+        public Link Toml { get; set; }
+
+        public AssetResponseLinks(Link toml)
+        {
+            Toml = toml;
+        }
+    }
+}
